requirements.txt
# Django and REST API
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2

# Database
psycopg2-binary==2.9.6
sqlparse==0.4.4

# Image Processing
Pillow==9.5.0

# Async Tasks
celery==5.2.7
redis==4.5.5

# AWS S3
boto3==1.26.137
django-storages==1.14.2

# Payment Integration
stripe==5.4.0
requests==2.31.0

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3

# Email
django-anymail==10.0

# API Documentation
drf-spectacular==0.26.2

# Testing
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.2.1

# Code Quality
black==23.3.0
flake8==6.0.0
isort==5.12.0
pylint==2.17.4

# Security
django-cors-headers==4.0.0
cryptography==41.0.0

# Production
gunicorn==20.1.0
whitenoise==6.4.0

# Monitoring
sentry-sdk==1.24.0
